fix(#1444): change entry to use asynchronous import
